[M3] Switched to using viewmodel.py to store view models
