+++ templates/index.html
@@ -48,3 +48,4 @@
 
   </div>
 {% endblock %}
+
